# bin/xbps-uhelper: allow multiple arguments for many actions