@classabbyamp 2023-02-09
> Useful!

would this be helpful for any xtools? I see `xpkg` uses sed at the moment